[PY-WEB-7]: Added login page
