+++ lab4/Khalabarchuk_Project/app/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask
+
+
+def enumerate_filter(iterable):
+    return enumerate(iterable)
+
+
+app = Flask(__name__)
+app.secret_key = b"secret"
+
+app.jinja_env.filters['enumerate'] = enumerate_filter
+
+
+from app import views
