+++ lab4/Khalabarchuk_Project/app/static/css/styles.css
@@ -8,13 +8,22 @@
     --black-color: #151515;
     --semi-black-color: rgba(21, 21, 21, 0.5);
 
+    --blue-color: rgb(47 184 255);
+    --blue-second-color: rgb(23 210 255);
+
+    --grean-color: #0ab930;
+    --grean-second-color: #2ac44d;
+
     --select-color: #ffe700;
     --select-hover-color: #ffe700;
 
     --white-gradient: linear-gradient(317deg, var(--white-color), var(--white-second-color));
-    --blue-gradient: linear-gradient(123deg, rgb(47 184 255) 39.22%, rgb(23 210 255) 90.31%);
+    --white-gradient-revers: linear-gradient(-317deg, var(--white-color), var(--white-second-color));
+    --blue-gradient: linear-gradient(123deg, var(--blue-color) 39.22%, var(--blue-second-color) 90.31%);
     --purple-gradient: linear-gradient(123deg, rgb(115, 33, 255) 39.22%, rgb(139, 84, 229) 90.31%);
     --light-purple-gradient: linear-gradient(123deg, rgb(139, 84, 229) 39.22%, rgb(169, 124, 245) 90.31%);
+
+    color: var(--black-color);
 }
 
 div, ul, body{
@@ -134,6 +143,55 @@ body{
 
 .default-button:active{
     transform: scale(98%);
+}
+
+.default-input{
+    max-width: 15vw;
+    width: 100%;
+
+    margin: 3px 0;
+    padding: 3px;
+
+    border-radius: 3px;
+    border: none;
+    border-bottom: 2px solid var(--blue-color);
+
+    font-size: .8rem;
+    font-weight: 600;
+
+    outline: none;
+
+    transition: ease .15s;
+}
+
+.default-input:focus{
+    scale: 1.2;
+
+    border-bottom: 2px solid var(--blue-second-color);
+}
+
+.default-input:hover:not(:focus) {
+    scale: 1.05;
+}
+
+.default-submit{
+    margin: 10px;
+    padding: 5px 15px;
+
+    border-radius: 10px;
+
+    font-size: 1.2rem;
+    font-weight: 600;
+
+    border: 2px solid var(--grean-color);
+
+    transition: ease .1s;
+}
+
+.default-submit:hover{
+    scale: 1.1;
+
+    border: 2px solid var(--grean-second-color);
 }
 
 footer{
