+++ lab4/Khalabarchuk_Project/app/static/images/projectWSHelper.png
